refactor: Extract a crate for shared data types in the new design
