+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "client_util",
     "compactor",
     "data_types",
+    "data_types2",
     "datafusion",
     "datafusion_util",
     "db",
