+++ data_types2/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "data_types2"
+version = "0.1.0"
+edition = "2021"
+description = "Shared data types in the Iox NG architecture"
+
+[dependencies]
+influxdb_line_protocol = { path = "../influxdb_line_protocol" }
+mutable_batch = { path = "../mutable_batch" }
+schema = { path = "../schema" }
+sqlx = { version = "0.5", features = [ "runtime-tokio-native-tls" , "postgres", "uuid" ] }
+uuid = { version = "0.8", features = ["v4"] }
+workspace-hack = { path = "../workspace-hack"}
