+++ data_types2/src/lib.rs
@@ -1,0 +1,638 @@
+//! Shared data types in the Iox NG architecture
+
+#![warn(
+    missing_copy_implementations,
+    missing_debug_implementations,
+    missing_docs,
+    clippy::explicit_iter_loop,
+    clippy::future_not_send,
+    clippy::use_self,
+    clippy::clone_on_ref_ptr
+)]
+
+use influxdb_line_protocol::FieldValue;
+use schema::{builder::SchemaBuilder, InfluxColumnType, InfluxFieldType, Schema};
+use std::{
+    collections::BTreeMap,
+    convert::TryFrom,
+    fmt::{Debug, Formatter},
+};
+use uuid::Uuid;
+
+/// Unique ID for a `Namespace`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct NamespaceId(i32);
+
+#[allow(missing_docs)]
+impl NamespaceId {
+    pub fn new(v: i32) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i32 {
+        self.0
+    }
+}
+
+impl std::fmt::Display for NamespaceId {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Unique ID for a `KafkaTopic`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct KafkaTopicId(i32);
+
+#[allow(missing_docs)]
+impl KafkaTopicId {
+    pub fn new(v: i32) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i32 {
+        self.0
+    }
+}
+
+impl std::fmt::Display for KafkaTopicId {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Unique ID for a `QueryPool`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct QueryPoolId(i16);
+
+#[allow(missing_docs)]
+impl QueryPoolId {
+    pub fn new(v: i16) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i16 {
+        self.0
+    }
+}
+
+/// Unique ID for a `Table`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct TableId(i32);
+
+#[allow(missing_docs)]
+impl TableId {
+    pub fn new(v: i32) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i32 {
+        self.0
+    }
+}
+
+impl std::fmt::Display for TableId {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Unique ID for a `Column`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct ColumnId(i32);
+
+#[allow(missing_docs)]
+impl ColumnId {
+    pub fn new(v: i32) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i32 {
+        self.0
+    }
+}
+
+/// Unique ID for a `Sequencer`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct SequencerId(i16);
+
+#[allow(missing_docs)]
+impl SequencerId {
+    pub fn new(v: i16) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i16 {
+        self.0
+    }
+}
+
+impl std::fmt::Display for SequencerId {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// The kafka partition identifier. This is in the actual Kafka cluster.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct KafkaPartition(i32);
+
+#[allow(missing_docs)]
+impl KafkaPartition {
+    pub fn new(v: i32) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i32 {
+        self.0
+    }
+}
+
+impl std::fmt::Display for KafkaPartition {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Unique ID for a `Partition`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct PartitionId(i64);
+
+#[allow(missing_docs)]
+impl PartitionId {
+    pub fn new(v: i64) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i64 {
+        self.0
+    }
+}
+
+impl std::fmt::Display for PartitionId {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Unique ID for a `Tombstone`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct TombstoneId(i64);
+
+#[allow(missing_docs)]
+impl TombstoneId {
+    pub fn new(v: i64) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i64 {
+        self.0
+    }
+}
+
+/// A sequence number from a `Sequencer` (kafka partition)
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct SequenceNumber(i64);
+
+#[allow(missing_docs)]
+impl SequenceNumber {
+    pub fn new(v: i64) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i64 {
+        self.0
+    }
+}
+
+/// A time in nanoseconds from epoch
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct Timestamp(i64);
+
+#[allow(missing_docs)]
+impl Timestamp {
+    pub fn new(v: i64) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i64 {
+        self.0
+    }
+}
+
+/// Unique ID for a `ParquetFile`
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type)]
+#[sqlx(transparent)]
+pub struct ParquetFileId(i64);
+
+#[allow(missing_docs)]
+impl ParquetFileId {
+    pub fn new(v: i64) -> Self {
+        Self(v)
+    }
+    pub fn get(&self) -> i64 {
+        self.0
+    }
+}
+
+impl std::fmt::Display for ParquetFileId {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        // Use `self.number` to refer to each positional data point.
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Data object for a kafka topic
+#[derive(Debug, Clone, Eq, PartialEq, sqlx::FromRow)]
+pub struct KafkaTopic {
+    /// The id of the topic
+    pub id: KafkaTopicId,
+    /// The unique name of the topic
+    pub name: String,
+}
+
+/// Data object for a query pool
+#[derive(Debug, Clone, Eq, PartialEq, sqlx::FromRow)]
+pub struct QueryPool {
+    /// The id of the pool
+    pub id: QueryPoolId,
+    /// The unique name of the pool
+    pub name: String,
+}
+
+/// Data object for a namespace
+#[derive(Debug, Clone, Eq, PartialEq, sqlx::FromRow)]
+pub struct Namespace {
+    /// The id of the namespace
+    pub id: NamespaceId,
+    /// The unique name of the namespace
+    pub name: String,
+    /// The retention duration as a string. 'inf' or not present represents infinite duration (i.e. never drop data).
+    #[sqlx(default)]
+    pub retention_duration: Option<String>,
+    /// The kafka topic that writes to this namespace will land in
+    pub kafka_topic_id: KafkaTopicId,
+    /// The query pool assigned to answer queries for this namespace
+    pub query_pool_id: QueryPoolId,
+    /// The maximum number of tables that can exist in this namespace
+    pub max_tables: i32,
+    /// The maximum number of columns per table in this namespace
+    pub max_columns_per_table: i32,
+}
+
+/// Schema collection for a namespace. This is an in-memory object useful for a schema
+/// cache.
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct NamespaceSchema {
+    /// the namespace id
+    pub id: NamespaceId,
+    /// the kafka topic this namespace gets data written to
+    pub kafka_topic_id: KafkaTopicId,
+    /// the query pool assigned to answer queries for this namespace
+    pub query_pool_id: QueryPoolId,
+    /// the tables in the namespace by name
+    pub tables: BTreeMap<String, TableSchema>,
+}
+
+impl NamespaceSchema {
+    /// Create a new `NamespaceSchema`
+    pub fn new(id: NamespaceId, kafka_topic_id: KafkaTopicId, query_pool_id: QueryPoolId) -> Self {
+        Self {
+            id,
+            tables: BTreeMap::new(),
+            kafka_topic_id,
+            query_pool_id,
+        }
+    }
+}
+
+/// Data object for a table
+#[derive(Debug, Clone, sqlx::FromRow, Eq, PartialEq)]
+pub struct Table {
+    /// The id of the table
+    pub id: TableId,
+    /// The namespace id that the table is in
+    pub namespace_id: NamespaceId,
+    /// The name of the table, which is unique within the associated namespace
+    pub name: String,
+}
+
+/// Column definitions for a table
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct TableSchema {
+    /// the table id
+    pub id: TableId,
+    /// the table's columns by their name
+    pub columns: BTreeMap<String, ColumnSchema>,
+}
+
+impl TableSchema {
+    /// Initialize new `TableSchema`
+    pub fn new(id: TableId) -> Self {
+        Self {
+            id,
+            columns: BTreeMap::new(),
+        }
+    }
+
+    /// Add `col` to this table schema.
+    ///
+    /// # Panics
+    ///
+    /// This method panics if a column of the same name already exists in
+    /// `self`, or the provided [`Column`] cannot be converted into a valid
+    /// [`ColumnSchema`].
+    pub fn add_column(&mut self, col: &Column) {
+        let old = self.columns.insert(
+            col.name.clone(),
+            ColumnSchema::try_from(col).expect("column is invalid"),
+        );
+        assert!(old.is_none());
+    }
+}
+
+/// Data object for a column
+#[derive(Debug, Clone, sqlx::FromRow, Eq, PartialEq)]
+pub struct Column {
+    /// the column id
+    pub id: ColumnId,
+    /// the table id the column is in
+    pub table_id: TableId,
+    /// the name of the column, which is unique in the table
+    pub name: String,
+    /// the logical type of the column
+    pub column_type: i16,
+}
+
+impl Column {
+    /// returns true if the column type is a tag
+    pub fn is_tag(&self) -> bool {
+        self.column_type == ColumnType::Tag as i16
+    }
+
+    /// returns true if the column type matches the line protocol field value type
+    pub fn matches_field_type(&self, field_value: &FieldValue) -> bool {
+        match field_value {
+            FieldValue::I64(_) => self.column_type == ColumnType::I64 as i16,
+            FieldValue::U64(_) => self.column_type == ColumnType::U64 as i16,
+            FieldValue::F64(_) => self.column_type == ColumnType::F64 as i16,
+            FieldValue::String(_) => self.column_type == ColumnType::String as i16,
+            FieldValue::Boolean(_) => self.column_type == ColumnType::Bool as i16,
+        }
+    }
+}
+
+/// The column id and its type for a column
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub struct ColumnSchema {
+    /// the column id
+    pub id: ColumnId,
+    /// the column type
+    pub column_type: ColumnType,
+}
+
+impl ColumnSchema {
+    /// returns true if the column is a tag
+    pub fn is_tag(&self) -> bool {
+        self.column_type == ColumnType::Tag
+    }
+
+    /// returns true if the column matches the line protocol field value type
+    pub fn matches_field_type(&self, field_value: &FieldValue) -> bool {
+        matches!(
+            (field_value, self.column_type),
+            (FieldValue::I64(_), ColumnType::I64)
+                | (FieldValue::U64(_), ColumnType::U64)
+                | (FieldValue::F64(_), ColumnType::F64)
+                | (FieldValue::String(_), ColumnType::String)
+                | (FieldValue::Boolean(_), ColumnType::Bool)
+        )
+    }
+
+    /// Returns true if `mb_column` is of the same type as `self`.
+    pub fn matches_type(&self, mb_column: &mutable_batch::column::Column) -> bool {
+        self.column_type == mb_column.influx_type()
+    }
+}
+
+impl TryFrom<&Column> for ColumnSchema {
+    type Error = Box<dyn std::error::Error>;
+
+    fn try_from(c: &Column) -> Result<Self, Self::Error> {
+        Ok(Self {
+            id: c.id,
+            column_type: ColumnType::try_from(c.column_type)?,
+        })
+    }
+}
+
+/// The column data type
+#[allow(missing_docs)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub enum ColumnType {
+    I64 = 1,
+    U64 = 2,
+    F64 = 3,
+    Bool = 4,
+    String = 5,
+    Time = 6,
+    Tag = 7,
+}
+
+impl ColumnType {
+    /// the short string description of the type
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            ColumnType::I64 => "i64",
+            ColumnType::U64 => "u64",
+            ColumnType::F64 => "f64",
+            ColumnType::Bool => "bool",
+            ColumnType::String => "string",
+            ColumnType::Time => "time",
+            ColumnType::Tag => "tag",
+        }
+    }
+}
+
+impl std::fmt::Display for ColumnType {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let s = self.as_str();
+
+        write!(f, "{}", s)
+    }
+}
+
+impl TryFrom<i16> for ColumnType {
+    type Error = Box<dyn std::error::Error>;
+
+    fn try_from(value: i16) -> Result<Self, Self::Error> {
+        match value {
+            x if x == Self::I64 as i16 => Ok(Self::I64),
+            x if x == Self::U64 as i16 => Ok(Self::U64),
+            x if x == Self::F64 as i16 => Ok(Self::F64),
+            x if x == Self::Bool as i16 => Ok(Self::Bool),
+            x if x == Self::String as i16 => Ok(Self::String),
+            x if x == Self::Time as i16 => Ok(Self::Time),
+            x if x == Self::Tag as i16 => Ok(Self::Tag),
+            _ => Err("invalid column value".into()),
+        }
+    }
+}
+
+impl From<InfluxColumnType> for ColumnType {
+    fn from(value: InfluxColumnType) -> Self {
+        match value {
+            InfluxColumnType::Tag => Self::Tag,
+            InfluxColumnType::Field(InfluxFieldType::Float) => Self::F64,
+            InfluxColumnType::Field(InfluxFieldType::Integer) => Self::I64,
+            InfluxColumnType::Field(InfluxFieldType::UInteger) => Self::U64,
+            InfluxColumnType::Field(InfluxFieldType::String) => Self::String,
+            InfluxColumnType::Field(InfluxFieldType::Boolean) => Self::Bool,
+            InfluxColumnType::Timestamp => Self::Time,
+        }
+    }
+}
+
+impl From<ColumnType> for InfluxColumnType {
+    fn from(value: ColumnType) -> Self {
+        match value {
+            ColumnType::I64 => Self::Field(InfluxFieldType::Integer),
+            ColumnType::U64 => Self::Field(InfluxFieldType::UInteger),
+            ColumnType::F64 => Self::Field(InfluxFieldType::Float),
+            ColumnType::Bool => Self::Field(InfluxFieldType::Boolean),
+            ColumnType::String => Self::Field(InfluxFieldType::String),
+            ColumnType::Time => Self::Timestamp,
+            ColumnType::Tag => Self::Tag,
+        }
+    }
+}
+
+impl TryFrom<TableSchema> for Schema {
+    type Error = schema::builder::Error;
+
+    fn try_from(value: TableSchema) -> Result<Self, Self::Error> {
+        let mut builder = SchemaBuilder::new();
+
+        for (column_name, column_schema) in &value.columns {
+            let t = InfluxColumnType::from(column_schema.column_type);
+            builder.influx_column(column_name, t);
+        }
+
+        builder.build()
+    }
+}
+
+impl PartialEq<InfluxColumnType> for ColumnType {
+    fn eq(&self, got: &InfluxColumnType) -> bool {
+        match self {
+            Self::I64 => matches!(got, InfluxColumnType::Field(InfluxFieldType::Integer)),
+            Self::U64 => matches!(got, InfluxColumnType::Field(InfluxFieldType::UInteger)),
+            Self::F64 => matches!(got, InfluxColumnType::Field(InfluxFieldType::Float)),
+            Self::Bool => matches!(got, InfluxColumnType::Field(InfluxFieldType::Boolean)),
+            Self::String => matches!(got, InfluxColumnType::Field(InfluxFieldType::String)),
+            Self::Time => matches!(got, InfluxColumnType::Timestamp),
+            Self::Tag => matches!(got, InfluxColumnType::Tag),
+        }
+    }
+}
+
+/// Returns the `ColumnType` for the passed in line protocol `FieldValue` type
+pub fn column_type_from_field(field_value: &FieldValue) -> ColumnType {
+    match field_value {
+        FieldValue::I64(_) => ColumnType::I64,
+        FieldValue::U64(_) => ColumnType::U64,
+        FieldValue::F64(_) => ColumnType::F64,
+        FieldValue::String(_) => ColumnType::String,
+        FieldValue::Boolean(_) => ColumnType::Bool,
+    }
+}
+
+/// Data object for a sequencer. Only one sequencer record can exist for a given
+/// kafka topic and partition (enforced via uniqueness constraint).
+#[derive(Debug, Copy, Clone, PartialEq, sqlx::FromRow)]
+pub struct Sequencer {
+    /// the id of the sequencer
+    pub id: SequencerId,
+    /// the topic the sequencer is reading from
+    pub kafka_topic_id: KafkaTopicId,
+    /// the kafka partition the sequencer is reading from
+    pub kafka_partition: KafkaPartition,
+    /// The minimum unpersisted sequence number. Because different tables
+    /// can be persisted at different times, it is possible some data has been persisted
+    /// with a higher sequence number than this. However, all data with a sequence number
+    /// lower than this must have been persisted to Parquet.
+    pub min_unpersisted_sequence_number: i64,
+}
+
+/// Data object for a partition. The combination of sequencer, table and key are unique (i.e. only
+/// one record can exist for each combo)
+#[derive(Debug, Clone, PartialEq, sqlx::FromRow)]
+pub struct Partition {
+    /// the id of the partition
+    pub id: PartitionId,
+    /// the sequencer the data in the partition arrived from
+    pub sequencer_id: SequencerId,
+    /// the table the partition is under
+    pub table_id: TableId,
+    /// the string key of the partition
+    pub partition_key: String,
+}
+
+/// Data object for a tombstone.
+#[derive(Debug, Clone, PartialEq, sqlx::FromRow)]
+pub struct Tombstone {
+    /// the id of the tombstone
+    pub id: TombstoneId,
+    /// the table the tombstone is associated with
+    pub table_id: TableId,
+    /// the sequencer the tombstone was sent through
+    pub sequencer_id: SequencerId,
+    /// the sequence nubmer assigned to the tombstone from the sequencer
+    pub sequence_number: SequenceNumber,
+    /// the min time (inclusive) that the delete applies to
+    pub min_time: Timestamp,
+    /// the max time (exclusive) that the delete applies to
+    pub max_time: Timestamp,
+    /// the full delete predicate
+    pub serialized_predicate: String,
+}
+
+/// Data for a parquet file reference in the catalog.
+#[derive(Debug, Clone, PartialEq, sqlx::FromRow)]
+pub struct ParquetFile {
+    /// the id of the file in the catalog
+    pub id: ParquetFileId,
+    /// the sequencer that sequenced writes that went into this file
+    pub sequencer_id: SequencerId,
+    /// the table
+    pub table_id: TableId,
+    /// the partition
+    pub partition_id: PartitionId,
+    /// the uuid used in the object store path for this file
+    pub object_store_id: Uuid,
+    /// the minimum sequence number from a record in this file
+    pub min_sequence_number: SequenceNumber,
+    /// the maximum sequence number from a record in this file
+    pub max_sequence_number: SequenceNumber,
+    /// the min timestamp of data in this file
+    pub min_time: Timestamp,
+    /// the max timestamp of data in this file
+    pub max_time: Timestamp,
+    /// flag to mark that this file should be deleted from object storage
+    pub to_delete: bool,
+    /// file size in bytes
+    pub file_size_bytes: i64,
+    /// thrift-encoded parquet metadata
+    pub parquet_metadata: Vec<u8>,
+    /// the number of rows of data in this file
+    pub row_count: i64,
+}
+
+/// Data for a processed tombstone reference in the catalog.
+#[derive(Debug, Copy, Clone, PartialEq, sqlx::FromRow)]
+pub struct ProcessedTombstone {
+    /// the id of the tombstone applied to the parquet file
+    pub tombstone_id: TombstoneId,
+    /// the id of the parquet file the tombstone was applied
+    pub parquet_file_id: ParquetFileId,
+}
