+++ ingester/Cargo.toml
@@ -14,6 +14,7 @@ base64 = "0.13"
 bytes = "1.0"
 datafusion = { path = "../datafusion" }
 data_types = { path = "../data_types" }
+data_types2 = { path = "../data_types2" }
 futures = "0.3"
 generated_types = { path = "../generated_types" }
 chrono = { version = "0.4", default-features = false }
