+++ ingester/src/querier_handler.rs
@@ -2,6 +2,7 @@
 
 use arrow::record_batch::RecordBatch;
 use arrow_util::util::merge_record_batches;
+use data_types2::SequencerId;
 use datafusion::{
     error::DataFusionError,
     physical_plan::{
@@ -10,7 +11,6 @@ use datafusion::{
         SendableRecordBatchStream,
     },
 };
-use iox_catalog::interface::SequencerId;
 use predicate::Predicate;
 use query::{
     exec::{Executor, ExecutorType},
