+++ iox_catalog/Cargo.toml
@@ -6,12 +6,11 @@ edition = "2021"
 
 [dependencies] # In alphabetical order
 async-trait = "0.1.42"
+data_types2 = { path = "../data_types2" }
 futures = "0.3"
-influxdb_line_protocol = { path = "../influxdb_line_protocol" }
 metric = { version = "0.1.0", path = "../metric" }
 mutable_batch = { path = "../mutable_batch" }
 observability_deps = { path = "../observability_deps" }
-schema = { path = "../schema" }
 snafu = "0.7"
 sqlx = { version = "0.5", features = [ "runtime-tokio-native-tls" , "postgres", "uuid" ] }
 sqlx-hotswap-pool = { path = "../sqlx-hotswap-pool" }
