+++ iox_catalog/src/postgres.rs
@@ -2,17 +2,19 @@
 
 use crate::{
     interface::{
-        sealed::TransactionFinalize, Catalog, Column, ColumnRepo, ColumnType, Error,
-        KafkaPartition, KafkaTopic, KafkaTopicId, KafkaTopicRepo, Namespace, NamespaceId,
-        NamespaceRepo, ParquetFile, ParquetFileId, ParquetFileRepo, Partition, PartitionId,
-        PartitionInfo, PartitionRepo, ProcessedTombstone, ProcessedTombstoneRepo, QueryPool,
-        QueryPoolId, QueryPoolRepo, RepoCollection, Result, SequenceNumber, Sequencer, SequencerId,
-        SequencerRepo, Table, TableId, TablePersistInfo, TableRepo, Timestamp, Tombstone,
-        TombstoneId, TombstoneRepo, Transaction,
+        sealed::TransactionFinalize, Catalog, ColumnRepo, Error, KafkaTopicRepo, NamespaceRepo,
+        ParquetFileRepo, PartitionInfo, PartitionRepo, ProcessedTombstoneRepo, QueryPoolRepo,
+        RepoCollection, Result, SequencerRepo, TablePersistInfo, TableRepo, TombstoneRepo,
+        Transaction,
     },
     metrics::MetricDecorator,
 };
 use async_trait::async_trait;
+use data_types2::{
+    Column, ColumnType, KafkaPartition, KafkaTopic, KafkaTopicId, Namespace, NamespaceId,
+    ParquetFile, ParquetFileId, Partition, PartitionId, ProcessedTombstone, QueryPool, QueryPoolId,
+    SequenceNumber, Sequencer, SequencerId, Table, TableId, Timestamp, Tombstone, TombstoneId,
+};
 use observability_deps::tracing::{info, warn};
 use sqlx::{migrate::Migrator, postgres::PgPoolOptions, Acquire, Executor, Postgres, Row};
 use sqlx_hotswap_pool::HotSwapPool;
