+++ iox_object_store/Cargo.toml
@@ -7,8 +7,8 @@ description = "IOx-specific semantics wrapping the general-purpose object store 
 [dependencies]
 bytes = "1.0"
 data_types = { path = "../data_types" }
+data_types2 = { path = "../data_types2" }
 futures = "0.3"
-iox_catalog = { path = "../iox_catalog" }
 object_store = { path = "../object_store" }
 observability_deps = { path = "../observability_deps" }
 snafu = "0.7"
