+++ querier/Cargo.toml
@@ -11,6 +11,7 @@ backoff = { path = "../backoff" }
 bytes = "1.0"
 client_util = { path = "../client_util" }
 data_types = { path = "../data_types" }
+data_types2 = { path = "../data_types2" }
 datafusion = { path = "../datafusion" }
 db = { path = "../db" }
 futures = "0.3"
